==> dev/22_predicting_persistence.Rmd <==
---
title: "Predict persistence"
author: "Chinmay Deval"
date: "2/14/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(tidymodels)
library(textrecipes)
library(finetune)
library(vip)
```

```{r, echo=FALSE, message=FALSE,}
fields= qs::qread("D:/OneDrive - University of Idaho/MagicValleyData/stats_merged_shp/new_stats/field_boundaries_with_all_stats.qs")%>% as.data.frame() %>% dplyr::select(-geometry)


fields_with_one_crop = qs::qread("D:/OneDrive - University of Idaho/MagicValleyData/CropScape_data_for_magicvalley/Common_land_cover_category/dominant_crop_cover_ts_shp/one_dominant_crop_over_entire_ts.qs")%>%
  as.data.frame()%>%dplyr::select("FID","Crop") 


fields_merged = dplyr::left_join(fields, fields_with_one_crop, by=c("FID"))%>%   dplyr::mutate_if(sapply(., is.character), as.factor)

fields_merged_all = fields_merged %>%as.data.frame()%>%
  dplyr::select(-c(FID, sndc_mn, sndc_mx, sndc_df,
                   Cr_2005, Cr_2007, Cr_2008, Cr_2009,
                   Cr_2010,Cr_2011,
                Cr_2012,Cr_2013,Cr_2014,Cr_2015,Cr_2016,Cr_2017,
                Cr_2018,Cr_2019, Cr_2020, Owner))%>%na.omit()%>% 
    dplyr::mutate(IRRI_ST = stringr::str_remove(IRRI_ST, "-"),
                  WR_class = stringr::str_remove(WR_class, "_WaterRight"),
                  RDR_classs = stringr::str_remove(RDR_classs, "_"),
                  Year = lubridate::year(PriorityDa),
                  IRRI_ST = as.factor(IRRI_ST),
                  WR_class = as.factor(WR_class),
                  RDR_classs = as.factor(RDR_classs))%>% dplyr::select(-PriorityDa)

```

```{r, echo=FALSE, message=FALSE,}
skimr::skim(fields_merged_all)
```

```{r}
fields_merged_all= fields_merged_all %>% dplyr::mutate(Prsstn_cls= dplyr::case_when(Prsstn_ == 0 ~ "No",
                                                     dplyr::between(Prsstn_, 0.001,30)~ "low",dplyr::between(Prsstn_, 30.0001,60)~ "Medium",
                                                     Prsstn_ > 60~ "High" ))
```


```{r, echo=FALSE, message=FALSE,warning=FALSE}
set.seed(123)

fields_split = fields_merged_all %>% initial_split(strata = Prsstn_)

fields_train = training(fields_split)
fields_test = testing(fields_split)



# crossvalidation folds
# field_folds =vfold_cv(fields_train, strata= Prsstn_)
# 
set.seed(789)
val_set <- validation_split(fields_train, 
                            strata = Prsstn_, 
                            prop = 0.80)
val_set

```



```{r, echo=FALSE, message=FALSE,warning=FALSE}
#pre-processing and feature engineering

fields_rec =recipe(Prsstn_ ~ ., data = fields_train)%>%
  step_tokenize(c("Crop", "IRRI_ST", "source","WR_class", "RDR_classs"))%>%
  step_tf(c("Crop", "IRRI_ST", "source","WR_class", "RDR_classs"))

fields_prep = prep(fields_rec)


juiced = juice(fields_prep)
```

```{r, echo=FALSE, message=FALSE,warning=FALSE}
fields_spec <- rand_forest(
  mtry = tune(),
  trees = 1000,
  min_n = tune()
) %>%
  set_mode("regression") %>%
  set_engine("ranger")
```

```{r, echo=FALSE, message=FALSE,warning=FALSE}
tune_wf <- workflow() %>%
  add_recipe(fields_rec) %>%
  add_model(fields_spec)
```

```{r}
# train hyper parameters
set.seed(234)
trees_folds <- vfold_cv(fields_train)


doParallel::registerDoParallel()

set.seed(345)
tune_res <- tune_grid(
  tune_wf,
  resamples = trees_folds,
  grid = 20
)
doParallel::stopImplicitCluster()
tune_res
```

```{r}
tune_res %>%
  collect_metrics() %>%
  filter(.metric == "rmse") %>%
  select(mean, min_n, mtry) %>%
  pivot_longer(min_n:mtry,
    values_to = "value",
    names_to = "parameter"
  ) %>%
  ggplot(aes(value, mean, color = parameter)) +
  geom_point(show.legend = FALSE) +
  facet_wrap(~parameter, scales = "free_x") +
  labs(x = NULL, y = "RMSE")
```

```{r}

rf_grid <- grid_regular(
  mtry(range = c(8, 25)),
  min_n(range = c(15, 40)),
  levels = 5
)

rf_grid
```


```{r}
doParallel::registerDoParallel()
set.seed(456)
regular_res <- tune_grid(
  tune_wf,
  resamples = trees_folds,
  grid = rf_grid
)
doParallel::stopImplicitCluster()
regular_res
```

```{r}
regular_res %>%
  collect_metrics() %>%
  filter(.metric == "rmse") %>%
  mutate(min_n = factor(min_n)) %>%
  ggplot(aes(mtry, mean, color = min_n)) +
  geom_line(alpha = 0.5, size = 1.5) +
  geom_point() +
  labs(y = "RMSE")
```


```{r}
best_rmse <- select_best(regular_res, "rmse")
doParallel::registerDoParallel()
final_rf <- finalize_model(
  fields_spec,
  best_rmse
)
doParallel::stopImplicitCluster()
foreach::registerDoSEQ()
final_rf
```


```{r}

final_rf %>%
  set_engine("ranger", importance = "permutation") %>%
  fit(Prsstn_ ~ .,
    data = juice(fields_prep)
  ) %>%
  vip(geom = "point")
```


```{r}
final_wf <- workflow() %>%
  add_recipe(fields_rec) %>%
  add_model(final_rf)

final_res <- final_wf %>%
  last_fit(fields_split)

final_res %>%
  collect_metrics()
```


```{r, echo=FALSE, message=FALSE,warning=FALSE}
# xgb_xpec = boost_tree(
#   trees = tune(),
#   mtry = tune(),
#   min_n = tune(),
#   learn_rate = 0.01
# )%>%
#   set_engine("xgboost")%>%
#   set_mode("regression")
# 
# xgb_wf = workflow(fields_rec, xgb_xpec)
```

```{r, echo=FALSE, message=FALSE,warning=FALSE}


# doParallel::registerDoParallel()
# set.seed(652)
# 
# xgb_fields_rs = tune_race_anova(
#   xgb_wf,
#   field_folds,
#   grid = 20,
#   control=control_race(verbose_elim = TRUE)
# )
# 
# xgb_fields_rs
```

```{r, echo=FALSE, message=FALSE,warning=FALSE}

```

```{r, echo=FALSE, message=FALSE,warning=FALSE}

```

```{r, echo=FALSE, message=FALSE,warning=FALSE}

```


```{r, echo=FALSE, message=FALSE,warning=FALSE}

```
